move nightmare to k command, remove massing
