+++ src/extendables/User/Minion.ts
@@ -363,8 +363,9 @@ export default class extends Extendable {
 
 			case 'Nightmare': {
 				const data = currentTask as NightmareActivityTaskOptions;
-
-				return `${this.minionName} is currently killing The Nightmare, with a party of ${data.users.length}. ${formattedDuration}`;
+				return `${this.minionName} is currently killing The Nightmare${
+					data.method === 'solo' ? 'solo' : 'in a team'
+				}. ${formattedDuration}`;
 			}
 
 			case 'AnimatedArmour': {
