+++ src/tasks/minions/minigames/nightmareActivity.ts
@@ -1,11 +1,10 @@
-import { noOp, percentChance } from 'e';
-import { KlasaUser, Task } from 'klasa';
+import { percentChance } from 'e';
+import { Task } from 'klasa';
 import { Bank, Misc } from 'oldschooljs';
 
-import { BitField, Emoji, NIGHTMARE_ID, PHOSANI_NIGHTMARE_ID } from '../../../lib/constants';
+import { BitField, NIGHTMARE_ID, PHOSANI_NIGHTMARE_ID } from '../../../lib/constants';
 import { addMonsterXP } from '../../../lib/minions/functions';
 import announceLoot from '../../../lib/minions/functions/announceLoot';
-import isImportantItemForMonster from '../../../lib/minions/functions/isImportantItemForMonster';
 import { trackLoot } from '../../../lib/settings/prisma';
 import { NightmareActivityTaskOptions } from '../../../lib/types/minions';
 import { randomVariation } from '../../../lib/util';
@@ -14,163 +13,95 @@ import { handleTripFinish } from '../../../lib/util/handleTripFinish';
 import { sendToChannelID } from '../../../lib/util/webhook';
 import { NightmareMonster } from './../../../lib/minions/data/killableMonsters/index';
 
-interface NightmareUser {
-	user: KlasaUser;
-	chanceOfDeath: number;
-	damageDone: number;
-}
-
 const RawNightmare = Misc.Nightmare;
 
 export default class extends Task {
 	async run(data: NightmareActivityTaskOptions) {
-		const { channelID, leader, users, quantity, duration, isPhosani = false } = data;
+		const { channelID, quantity, duration, isPhosani = false, userID, method } = data;
 
 		const monsterID = isPhosani ? PHOSANI_NIGHTMARE_ID : NightmareMonster.id;
 		const monsterName = isPhosani ? "Phosani's Nightmare" : 'Nightmare';
+		const user = await this.client.fetchUser(userID);
+		const team = method === 'solo' ? [user.id] : [user.id, '1', '2', '3'];
 
-		const teamsLoot: { [key: string]: Bank } = {};
-		const teamsPreviousCL: { [key: string]: Bank } = {};
-		const kcAmounts: { [key: string]: number } = {};
-
-		const parsedUsers: NightmareUser[] = [];
-		const totalLoot = new Bank();
-
-		// For each user in the party, calculate their damage and death chance.
-		for (const id of users) {
-			const user = await this.client.fetchUser(id).catch(noOp);
-			if (!user) continue;
-			const [data] = getNightmareGearStats(user, users, isPhosani);
-			parsedUsers.push({ ...data, user });
-			teamsLoot[id] = new Bank();
-		}
-
-		// Store total amount of deaths
-		const deaths: Record<string, number> = {};
+		const [userStats] = getNightmareGearStats(user, team, isPhosani);
+		const parsedUsers = team.map(id => ({ ...userStats, id }));
+		const userLoot = new Bank();
+		let kc = 0;
+		let deaths = 0;
 
 		for (let i = 0; i < quantity; i++) {
 			const _loot = RawNightmare.kill({
 				team: parsedUsers.map(user => ({
-					id: user.user.id,
-					damageDone: users.length === 1 ? 2400 : randomVariation(user.damageDone, 5)
+					id: user.id,
+					damageDone: team.length === 1 ? 2400 : randomVariation(user.damageDone, 5)
 				})),
 				isPhosani
 			});
 
-			const loot: Record<string, Bank> = {};
-			for (const [id, itemBank] of Object.entries(_loot)) {
-				loot[id] = new Bank(itemBank);
-			}
-
-			for (const { user, chanceOfDeath } of parsedUsers) {
-				kcAmounts[user.id] = Boolean(kcAmounts[user.id]) ? ++kcAmounts[user.id] : 1;
-				const died = percentChance(chanceOfDeath);
-				if (died) {
-					deaths[user.id] = deaths[user.id] ? ++deaths[user.id] : 1;
-					kcAmounts[user.id]--;
-				} else {
-					if (user.owns('Slepey tablet') || user.bitfield.includes(BitField.HasSlepeyTablet)) {
-						loot[user.id].remove('Slepey tablet', loot[user.id].amount('Slepey tablet'));
-					}
-					loot[user.id].bank;
-					teamsLoot[user.id].add(loot[user.id]);
-				}
+			const died = percentChance(userStats.chanceOfDeath);
+			if (died) {
+				deaths++;
+			} else {
+				userLoot.add(_loot[user.id]);
+				kc++;
 			}
 		}
 
-		const leaderUser = await this.client.fetchUser(leader);
+		await addMonsterXP(user, {
+			monsterID: NIGHTMARE_ID,
+			quantity: Math.ceil(quantity / team.length),
+			duration,
+			isOnTask: false,
+			taskQuantity: null
+		});
 
-		let resultStr = `${leaderUser}, your party finished killing ${quantity}x ${NightmareMonster.name}!\n\n`;
-
-		for (const [userID, loot] of Object.entries(teamsLoot)) {
-			const user = await this.client.fetchUser(userID).catch(noOp);
-			if (!user) continue;
-			await addMonsterXP(user, {
-				monsterID: NIGHTMARE_ID,
-				quantity: Math.ceil(quantity / users.length),
-				duration,
-				isOnTask: false,
-				taskQuantity: null
-			});
-
-			// Fix purple items on solo kills
-			const { previousCL, itemsAdded } = await user.addItemsToBank({ items: loot, collectionLog: true });
-
-			totalLoot.add(itemsAdded);
-
-			// Only add previousCL for leader
-			if (user.id === leader) teamsPreviousCL[user.id] = previousCL;
-
-			const kcToAdd = kcAmounts[user.id];
-			if (kcToAdd) await user.incrementMonsterScore(monsterID, kcToAdd);
-			const purple = loot.items().some(([item]) => isImportantItemForMonster(item.id, NightmareMonster));
-
-			resultStr += `${purple ? Emoji.Purple : ''} **${user} received:** ||${loot}||\n`;
-
-			announceLoot({
-				user,
-				monsterID,
-				loot,
-				notifyDrops: NightmareMonster.notifyDrops,
-				team: {
-					leader: leaderUser,
-					lootRecipient: user,
-					size: users.length
-				}
-			});
+		if (user.owns('Slepey tablet') || user.bitfield.includes(BitField.HasSlepeyTablet)) {
+			userLoot.remove('Slepey tablet', userLoot.amount('Slepey tablet'));
 		}
+		// Fix purple items on solo kills
+		const { previousCL, itemsAdded } = await user.addItemsToBank({ items: userLoot, collectionLog: true });
 
-		// Show deaths in the result
-		const deathEntries = Object.entries(deaths);
-		if (deathEntries.length > 0) {
-			const deaths = [];
-			for (const [id, qty] of deathEntries) {
-				const user = await this.client.fetchUser(id).catch(noOp);
-				if (!user) continue;
-				deaths.push(`**${user.username}**: ${qty}x`);
-			}
-			resultStr += `\n**Deaths**: ${deaths.join(', ')}.`;
-		}
+		if (kc) await user.incrementMonsterScore(monsterID, kc);
+
+		announceLoot({
+			user,
+			monsterID,
+			loot: itemsAdded,
+			notifyDrops: NightmareMonster.notifyDrops
+		});
 
 		await trackLoot({
-			loot: totalLoot,
+			loot: itemsAdded,
 			id: monsterName,
 			type: 'Monster',
 			changeType: 'loot',
 			duration,
-			kc: quantity,
-			teamSize: users.length
+			kc: quantity
 		});
 
-		if (users.length > 1) {
-			sendToChannelID(channelID, { content: resultStr });
-		} else if (!kcAmounts[leader]) {
+		if (!kc) {
 			sendToChannelID(channelID, {
-				content: `${leaderUser}, ${leaderUser.minionName} died in all their attempts to kill the ${monsterName}, they apologize and promise to try harder next time.`
+				content: `${user}, ${user.minionName} died in all their attempts to kill the ${monsterName}, they apologize and promise to try harder next time.`
 			});
 		} else {
 			const { image } = await this.client.tasks
 				.get('bankImage')!
-				.generateBankImage(
-					teamsLoot[leader],
-					`${quantity}x Nightmare`,
-					true,
-					{ showNewCL: 1 },
-					leaderUser,
-					teamsPreviousCL[leader]
-				);
+				.generateBankImage(itemsAdded, `${quantity}x Nightmare`, true, { showNewCL: 1 }, user, previousCL);
 
-			const kc = leaderUser.getKC(monsterID);
+			const kc = user.getKC(monsterID);
 			handleTripFinish(
-				leaderUser,
+				user,
 				channelID,
-				`${leaderUser}, ${leaderUser.minionName} finished killing ${quantity} ${monsterName}, you died ${
-					deaths[leader] ?? 0
-				} times. Your ${monsterName} KC is now ${kc}.`,
-				['nightmare', [isPhosani ? 'phosani' : 'solo'], true],
+				`${user}, ${user.minionName} finished killing ${quantity} ${monsterName}, you died ${deaths} times. Your ${monsterName} KC is now ${kc}.`,
+				[
+					'k',
+					{ name: isPhosani ? 'phosani nightmare' : method === 'solo' ? 'solo nightmare' : 'mass nightmare' },
+					true
+				],
 				image!,
 				data,
-				totalLoot
+				itemsAdded
 			);
 		}
 	}
